+++ firmware/drivers/isis_antenna/isis_antenna_delay.c
@@ -43,7 +43,7 @@
 
 void isis_antenna_delay_s(uint8_t s)
 {
-    vTaskDelay(pdMS_TO_TICKS(1000U*s));
+    vTaskDelay(pdMS_TO_TICKS(1000U*(uint32_t)s));
 }
 
 void isis_antenna_delay_ms(uint16_t ms)
